Replace SliceExt with inherent [T] methods in libcore
